Improve book removal functionality
The book removal button now splices the corresponding object from
the library array. Because this messes with the index order of the
other objects, the books on the page are reloaded, and therefore
reassigned new data-index numbers. The previous method would have
left behind a mess of empty objects in the array.